+++ lib/openapi_first/schema_validation.rb
@@ -3,7 +3,6 @@
 require 'json_schemer'
 
 module OpenapiFirst
-  # Wraps JSONSchemer, mostly to add raw_schema method to test things :|
   class SchemaValidation
     attr_reader :raw_schema
 
@@ -14,12 +13,24 @@ module OpenapiFirst
       custom_keywords['readOnly'] = proc { |data| !data } if write
       @schemer = JSONSchemer.schema(
         schema,
-        keywords: custom_keywords
+        keywords: custom_keywords,
+        before_property_validation: proc do |data, property, property_schema, parent|
+          convert_nullable(data, property, property_schema, parent)
+        end
       )
     end
 
     def validate(input)
       @schemer.validate(input)
+    end
+
+    private
+
+    def convert_nullable(_data, _property, property_schema, _parent)
+      return unless property_schema.is_a?(Hash) && property_schema['nullable'] && property_schema['type']
+
+      property_schema['type'] = [*property_schema['type'], 'null']
+      property_schema.delete('nullable')
     end
   end
 end